[fix/#274] 푸시 알림 토큰 삭제 로직 위치 수정 (탈퇴시 삭제되도록 변경)
