+++ feature/mypage/src/main/java/com/napzak/market/mypage/withdraw/WithdrawViewModel.kt
@@ -5,6 +5,8 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.napzak.market.notification.repository.NotificationRepository
+import com.napzak.market.notification.usecase.DeletePushTokenUseCase
 import com.napzak.market.store.usecase.WithdrawUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableSharedFlow
@@ -16,6 +18,8 @@ import javax.inject.Inject
 @HiltViewModel
 internal class WithdrawViewModel @Inject constructor(
     private val withdrawUseCase: WithdrawUseCase,
+    private val deletePushTokenUseCase: DeletePushTokenUseCase,
+    private val notificationRepository: NotificationRepository,
 ) : ViewModel() {
     private val _sideEffect = MutableSharedFlow<WithdrawSideEffect>()
     val sideEffect = _sideEffect.asSharedFlow()
@@ -24,6 +28,18 @@ internal class WithdrawViewModel @Inject constructor(
     var withdrawDescription by mutableStateOf("")
     var isWithdrawing by mutableStateOf(false)
         private set
+
+    fun deletePushToken() {
+        viewModelScope.launch {
+            val pushToken = notificationRepository.getPushToken()
+            if (pushToken != null) {
+                deletePushTokenUseCase(pushToken)
+                    .onSuccess { notificationRepository.cleanPushToken() }
+                    .onFailure { Timber.e(it) }
+
+            }
+        }
+    }
 
     fun withdrawStore() {
         viewModelScope.launch {
